chore: split up core structs into separate files
